Vendor HEAD of protoeasy but stop using gb to build it for now
Downside is this means the build isn't consistently reproducible as
building protoeasy relies on libraries present in $GOPATH (i.e. outside
of this repository) but this allows skipping the lion/protobuf debacle
whilst still keeping protoeasy changes under this repository. We'll get
back to a standard setup once the Lion stuff has settled.