BREAD-1533: Refactors build script to be Rake-based
I went back and forth on this a bit, but I think Ruby's `make` (`rake`) makes a
lot of sense for this task. I like the ability to lightly use Ruby structs and
code, and leaving the door open for some images to build themselves in a custom
way by adding a Rakefile or Makefile to their own directory.

Original-Commit: 6e3b0b220d29655b19232ac0adad6acc69a93ad7
Original-Commit-URL: https://git.dev.pardot.com/Pardot/docker-library/commit/6e3b0b220d29655b19232ac0adad6acc69a93ad7