BREAD-1636: Renames artifactory_integration to internal_tools_integration
Since this VPC is already setup with everything we need to do a
pardot-atlassian -> pardotops transition, I think it makes sense to put GitHub
Enterprise here.

The idea here is that this is a no-op as far as resources on AWS, but just an
update in state. I'll use https://www.terraform.io/docs/commands/state/mv.html
to rename resources. **The result is that the plan for this change should be
empty.** I won't be applying anything.