BREAD-1913 /merge merges a PR if the only commits after the command are CI automerges
This PR makes sure that the `/merge` command still operates correctly if all that happened since the command was issued was a CI automerge from master.

If commits other than a CI automerge are pushed since the `/merge` command is issued, jib will reconfirm with the user about the intent to merge.