Upgrade to github.com/sr/operator@36fe5e03
The main change here is that services now need to opt-in to be exposed
as gRPC/chat commands. Encoding this type of configuration directly in
the proto files allows just running the same "protoc" command over all
proto files and code generation be controller there instead of at the
build script level.