Bring gRPC interceptor in house and drop Instrumenter altogether
Having it here makes things easier to modify and understand. I think
going forward I'd like to "dismentle" a lot of the operator package and
bring most of it under our source tree. The real value of operator is in
the protobuf generator code, which can remain as an external dependency.
However, all the abstractions around serving gRPC (logging, auth, ...)
and webhooks requests belong here I think. This is a litte bit a case of
premature abstraction.